frontend/requirements.txt
dash
flask
plotly
pandas
numpy
sqlalchemy
scipy
psycopg2-binary